% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GBload}
\alias{GBload}
\title{The electricity actual total load in Great Britain in year 2018}
\format{
A data frame with 17520 rows and 2 variables:
\describe{
\item{Date}{date indicating the delivery beginning of the electricity}
\item{Load}{actual electricity load in MW}
...
}
}
\source{
\url{https://transparency.entsoe.eu/}
}
\usage{
GBload
}
\description{
A dataset containing the electricity actual total load (MW) in
Great Britain in year 2018 presented in half-hour interval. Each data
point regards 30 minutes of electricity load starting at given time. The
data consists of both missing values and outliers.
}
\keyword{datasets}
