% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_modelled_data.R
\name{impute_modelled_data}
\alias{impute_modelled_data}
\title{Impute modelled missing time series data}
\usage{
impute_modelled_data(object, tau = NULL)
}
\arguments{
\item{object}{an object of class "tsrobprep" that is an output of functions
model_missing_data and handle_outliers.}

\item{tau}{the quantile(s) of the missing values to be imputed. tau should be
a subset of the quantile values present in the "tsrobprep" object. By default
all quantiles present in the object are used.}
}
\value{
A matrix or a list of matrices with imputed missing values or
outliers.
}
\description{
Returns a matrix or a list of matrices with imputed missing values or
outliers. As argument the function requires an object of class "tsrobprep"
and the quantiles to be imputed.
}
\examples{
model.miss <- model_missing_data(data = GBload[,-1], S = 48, tau = 0.5,
              no.of.last.indices.to.fix = dim(GBload)[1], consider.as.missing = 0,
              min.val = 0)
model.miss$estimated.models
model.miss$replaced.indices
new.GBload <- impute_modelled_data(model.miss)
}
\seealso{
\link[tsrobprep]{model_missing_data}, \link[tsrobprep]{handle_outliers},
\link[tsrobprep]{auto_data_cleaning}
}
